third_party/blink/web_tests/external/wpt/content-security-policy/sandbox/service-worker-sandbox.https-expected.txt

This is a testharness.js-based test.
[FAIL] Origin of service worker
  assert_equals: Origin should not be sandboxed expected "https://web-platform.test:8444" but got "null"
[FAIL] Response generated by service worker can be fetched as same-origin
  assert_equals: Origin should not be opaque expected "https://web-platform.test:8444" but got "null"
[FAIL] Origin used in fetch on service worker (mode: same-origin, with ACAOrigin)
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Origin used in fetch on service worker (mode: same-origin)
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Origin used in fetch on service worker (mode: cors)
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
Harness: the test ran to completion.
